Update README + renames.
