+++ README.md
@@ -48,26 +48,23 @@ pseudo-random number generator by George Marsaglia that has good statistical
 properties. See the [xorshift example](examples/example_xorshift.c).
 
 ## Hashing
-- The [FNV hash](https://tools.ietf.org/html/draft-eastlake-fnv-17) is fast
+- The [FNV hash](https://tools.ietf.org/html/draft-eastlake-fnv-17) is a fast
 hash function that maps variable length input to a fixed output
 ([example](examples/example_FNV_hash.c)). It can be used for [feature
 hashing](https://en.wikipedia.org/wiki/Feature_hashing).
 
-## Online statistics
+## Statistics
 - Welfords method computes mean and variance in a single pass. See the
 [example of Welford's method](examples/example_Welfords_method.c).
 
-## Feature extraction
+## Transformations
 - Fast Walsh-Hadamard transform (FWHT) implements the Walsh-Hardamard
 transform in O(n log n) time. The FWHT similar to the fast Fourier transform
 and the Haar transform. See the [FWHT example](examples/example_FWHT.c).
 
-## Regression
+## Passive-aggressive learning
 - [Online passive-aggressive (PA)](docs/crammer2006opa.pdf) regression solves a
 regression problem by only updating the model on prediction mistakes. 
-
-## Classification
-- Kernel passive-aggressive classification (TODO).
 
 # Other solutions for Tiny ML or Edge AI
 - [TensorFlow Lite](https://www.tensorflow.org/lite/)
