+++ examples/example_hash.c
