+++ examples/example_rng.c
