+++ examples/example_stats.c
