+++ tests/transform_test.c
@@ -1,4 +1,4 @@
-#include <fwht.h>
+#include "transform.h"
 #include <math.h>
 #include <stdint.h>
 #include <stdio.h>
